Properly support using an unknown class inside a function.
Remove RuntimeInstance, replace it with just a PepVariable,
and allow PepTypeMatchers to return a namespace from the
runtime_namespace() method, instead of every type providing
a runtime_instance() method.  Require that the type passed in
to PepVariable's constructor be a known type, i.e. an instance
of PepTypeMatcher, so that PepVariable can return a runtime
namespace when asked.